+++ c-v/src/components/Modal/AddCommandModal.jsx
@@ -3,7 +3,7 @@ import ModalCard from '../../styles/components/ModalCard';
 import ModalTitle from '../../styles/components/ModalTitle';
 import ConfirmButtons from '../../styles/components/ConfirmButtons';
 import Button from '../common/Button';
-import EditInput from './EditInput';
+import EditInput from './EditList/EditInput';
 import useAddCommand from '../../hooks/useAddCommand';
 import { useAddCommandModalStore } from '../stores/ModalStore';
